Corrigido problema no map de retorno do thingsboard

+++ src/main/java/com/ifsc/julio/javatcc/dto/DeviceTelemetryDayDTO.java
@@ -1,14 +1,13 @@
 package com.ifsc.julio.javatcc.dto;
 
 import lombok.*;
-import java.math.BigDecimal;
 import java.util.Date;
 
 @Getter @Setter
 @AllArgsConstructor @NoArgsConstructor
 @Builder
 public class DeviceTelemetryDayDTO {
-    private BigDecimal average;
+    private Double average;
     private Date day;
     private String key;
 }
