+++ src/main/java/com/ifsc/julio/javatcc/entity/DeviceTelemetryDayEntity.java
@@ -5,7 +5,6 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import lombok.*;
-import java.math.BigDecimal;
 import java.util.Date;
 import java.util.UUID;
 
@@ -22,5 +21,5 @@ public class DeviceTelemetryDayEntity {
 
     private String key;
     private Date date;
-    private BigDecimal value;
+    private Double value;
 }
