+++ src/main/java/com/ifsc/julio/javatcc/entity/DeviceTelemetryHourEntity.java
@@ -2,7 +2,6 @@ package com.ifsc.julio.javatcc.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import java.math.BigDecimal;
 import java.util.Date;
 import java.util.UUID;
 
@@ -21,5 +20,5 @@ public class DeviceTelemetryHourEntity {
 
     private String key;
     private Date date;
-    private BigDecimal value;
+    private Double value;
 }
